int (num?) to double

+++ lib/models/balance.dart
@@ -4,7 +4,7 @@ class Balance {
   Balance(this.amount);
 
   Balance.fromJson(Map<dynamic, dynamic> json)
-    : amount = json['amount'] as double;
+    : amount = json['amount'].toDouble() as double;
 
   Map<dynamic, dynamic> toJson() => <dynamic, dynamic>{
     'amount': amount,
